+++ graal-nodejs/mx.graal-nodejs/suite.py
@@ -2,7 +2,7 @@
 suite = {
   "mxversion" : "7.58.0",
   "name" : "graal-nodejs",
-  "version" : "26.0.0",
+  "version" : "25.1.0",
   "release" : False,
   "versionConflictResolution" : "latest",